Split up unit tests and improve naming
